// libs/mira-v1-ts/src/sdk/typegen/RemoveLiquidityScriptLoader.ts
export type IdentityInput = Enum<{
  Address: AddressInput;
  ContractId: ContractIdInput;
}>;
export type IdentityOutput = Enum<{
  Address: AddressOutput;
  ContractId: ContractIdOutput;
}>;

export type AddressInput = {bits: string};
export type AddressOutput = AddressInput;
export type AssetIdInput = {bits: string};
export type AssetIdOutput = AssetIdInput;
export type ContractIdInput = {bits: string};
export type ContractIdOutput = ContractIdInput;

export type RemoveLiquidityScriptLoaderInputs = [
  pool_id: [AssetIdInput, AssetIdInput, boolean],
  liquidity: BigNumberish,
  amount_0_min: BigNumberish,
  amount_1_min: BigNumberish,
  recipient: IdentityInput,
  deadline: BigNumberish,
];
export type RemoveLiquidityScriptLoaderOutput = [BN, BN];

export type RemoveLiquidityScriptLoaderConfigurables = Partial<{
  AMM_CONTRACT_ID: ContractIdInput;
}>;

const abi = {
  programType: "script",
  specVersion: "1",
  encodingVersion: "1",
  concreteTypes: [
    {
      type: "(struct std::asset_id::AssetId, struct std::asset_id::AssetId, bool)",
      concreteTypeId:
        "a95e1fcceb1451b8a76471f593f66c4a52ca04bde3c227c746ad7aaf988de5c6",
      metadataTypeId: 1,
    },
    {
      type: "(u64, u64)",
      concreteTypeId:
        "41bd1a98f0a59642d8f824c805b798a5f268d1f7d05808eb05c4189c493f1be0",
      metadataTypeId: 0,
    },
    {
      type: "enum std::identity::Identity",
      concreteTypeId:
        "ab7cd04e05be58e3fc15d424c2c4a57f824a2a2d97d67252440a3925ebdc1335",
      metadataTypeId: 4,
    },
    {
      type: "str",
      concreteTypeId:
        "8c25cb3686462e9a86d2883c5688a22fe738b0bbc85f458d2d2b5f3f667c6d5a",
    },
    {
      type: "struct std::contract_id::ContractId",
      concreteTypeId:
        "29c10735d33b5159f0c71ee1dbd17b36a3e69e41f00fab0d42e1bd9f428d8a54",
      metadataTypeId: 7,
    },
    {
      type: "u32",
      concreteTypeId:
        "d7649d428b9ff33d188ecbf38a7e4d8fd167fa01b2e10fe9a8f9308e52f1d7cc",
    },
    {
      type: "u64",
      concreteTypeId:
        "1506e6f44c1d6291cdf46395a8e573276a4fa79e8ace3fc891e092ef32d1b0a0",
    },
  ],
  metadataTypes: [
    {
      type: "(_, _)",
      metadataTypeId: 0,
      components: [
        {
          name: "__tuple_element",
          typeId:
            "1506e6f44c1d6291cdf46395a8e573276a4fa79e8ace3fc891e092ef32d1b0a0",
        },
        {
          name: "__tuple_element",
          typeId:
            "1506e6f44c1d6291cdf46395a8e573276a4fa79e8ace3fc891e092ef32d1b0a0",
        },
      ],
    },
    {
      type: "(_, _, _)",
      metadataTypeId: 1,
      components: [
        {
          name: "__tuple_element",
          typeId: 6,
        },
        {
          name: "__tuple_element",
          typeId: 6,
        },
        {
          name: "__tuple_element",
          typeId: 3,
        },
      ],
    },
    {
      type: "b256",
      metadataTypeId: 2,
    },
    {
      type: "bool",
      metadataTypeId: 3,
    },
    {
      type: "enum std::identity::Identity",
      metadataTypeId: 4,
      components: [
        {
          name: "Address",
          typeId: 5,
        },
        {
          name: "ContractId",
          typeId: 7,
        },
      ],
    },
    {
      type: "struct std::address::Address",
      metadataTypeId: 5,
      components: [
        {
          name: "bits",
          typeId: 2,
        },
      ],
    },
    {
      type: "struct std::asset_id::AssetId",
      metadataTypeId: 6,
      components: [
        {
          name: "bits",
          typeId: 2,
        },
      ],
    },
    {
      type: "struct std::contract_id::ContractId",
      metadataTypeId: 7,
      components: [
        {
          name: "bits",
          typeId: 2,
        },
      ],
    },
  ],
  functions: [
    {
      inputs: [
        {
          name: "pool_id",
          concreteTypeId:
            "a95e1fcceb1451b8a76471f593f66c4a52ca04bde3c227c746ad7aaf988de5c6",
        },
        {
          name: "liquidity",
          concreteTypeId:
            "1506e6f44c1d6291cdf46395a8e573276a4fa79e8ace3fc891e092ef32d1b0a0",
        },
        {
          name: "amount_0_min",
          concreteTypeId:
            "1506e6f44c1d6291cdf46395a8e573276a4fa79e8ace3fc891e092ef32d1b0a0",
        },
        {
          name: "amount_1_min",
          concreteTypeId:
            "1506e6f44c1d6291cdf46395a8e573276a4fa79e8ace3fc891e092ef32d1b0a0",
        },
        {
          name: "recipient",
          concreteTypeId:
            "ab7cd04e05be58e3fc15d424c2c4a57f824a2a2d97d67252440a3925ebdc1335",
        },
        {
          name: "deadline",
          concreteTypeId:
            "d7649d428b9ff33d188ecbf38a7e4d8fd167fa01b2e10fe9a8f9308e52f1d7cc",
        },
      ],
      name: "main",
      output:
        "41bd1a98f0a59642d8f824c805b798a5f268d1f7d05808eb05c4189c493f1be0",
      attributes: null,
    },
  ],
  loggedTypes: [
    {
      logId: "10098701174489624218",
      concreteTypeId:
        "8c25cb3686462e9a86d2883c5688a22fe738b0bbc85f458d2d2b5f3f667c6d5a",
    },
  ],
  messagesTypes: [],
  configurables: [
    {
      name: "AMM_CONTRACT_ID",
      concreteTypeId:
        "29c10735d33b5159f0c71ee1dbd17b36a3e69e41f00fab0d42e1bd9f428d8a54",
      offset: 88,
    },
  ],
};

const bytecode = decompressBytecode(
  "H4sIAAAAAAAAA5NyMGAIcGQwkHIJYNjlycBg5MDSCOQrxALZQJoDyG9ScBVmCHIVYPFyYWBo0zphtK15kfLCWzN+vIp221zq0Xr763S2qNlONVmL7qQtZ4CAA1AaL/gPBC6piSk5mXmpCgWJxcWpKQT1JJUW5YFoz7zi0rS0zOTM1LwShcTc/NK8ErD8GTBgYAKZDQI9qqfN2tz0ZkF4MFPEX0DpDxBaggNCC3VAaM4fADSL2yUYAQAA"
);

export class RemoveLiquidityScriptLoader extends Script<
  RemoveLiquidityScriptLoaderInputs,
  RemoveLiquidityScriptLoaderOutput
> {
  static readonly abi = abi;
  static readonly bytecode = bytecode;

  constructor(wallet: Account) {
    super(bytecode, abi, wallet);
  }
}