// libs/mira-v1-ts/src/sdk/typegen/SwapExactInputScriptLoader.ts
export type IdentityInput = Enum<{ Address: AddressInput, ContractId: ContractIdInput }>;
export type IdentityOutput = Enum<{ Address: AddressOutput, ContractId: ContractIdOutput }>;

export type AddressInput = { bits: string };
export type AddressOutput = AddressInput;
export type AssetIdInput = { bits: string };
export type AssetIdOutput = AssetIdInput;
export type ContractIdInput = { bits: string };
export type ContractIdOutput = ContractIdInput;

export type SwapExactInputScriptLoaderInputs = [amount_in: BigNumberish, asset_in: AssetIdInput, amount_out_min: BigNumberish, pools: Vec<[AssetIdInput, AssetIdInput, boolean]>, recipient: IdentityInput, deadline: BigNumberish];
export type SwapExactInputScriptLoaderOutput = Vec<[BN, AssetIdOutput]>;
// ...
const bytecode = decompressBytecode('H4sIAAAAAAAAA5NyMGAIcGQwkHIJYNjlycBg5MDSCOQrxALZQJoDyG9ScBVmCHIVYPFyYWDgOqfMztL5bF35lcl9Vlfsy6zk0gI08sqdZm+uiFFZfu8EAxgwzYDQ+MF/IHBJTUzJycxLVShILC5OTWEIyi8tSS2yUvD0C3P08XSJD3AM8QCpTUtNLSbGzIL8/Jz43NSSxJTEkkQgPwDIV8jLL1EoKEotTs0r8cwrLk1Ly0zOBLIVgHYVlJYoJObml+aVQPQXlycWwNwGAmfAgIEJwmNg7FE9bdbmpjcLyicEGHFJ8D7Ytnl5CkH9zPil1QUgtO0UKD0HSq+B0FZHILTlAQhtkAChtRZA6QaoOVCXqN6B0II7oDTUPMEcKA2OCyCtAqEFvkDpECitAaH5oeL8FyA07xko7QOhefZAaaj9PFBzeaD6eXggNPcbCM31AEJzguOGgYED6l/2K1DaAkKz9UBpqP1sUPtZofaz2QAAT+4ohvACAAA=');

export class SwapExactInputScriptLoader extends Script<SwapExactInputScriptLoaderInputs, SwapExactInputScriptLoaderOutput> {

  static readonly abi = abi;
  static readonly bytecode = bytecode;

  constructor(wallet: Account) {
    super(bytecode, abi, wallet);
  }
}