// libs/mira-v1-ts/src/sdk/typegen/SwapExactOutputScriptLoader.ts
export type IdentityInput = Enum<{
  Address: AddressInput;
  ContractId: ContractIdInput;
}>;
export type IdentityOutput = Enum<{
  Address: AddressOutput;
  ContractId: ContractIdOutput;
}>;

export type AddressInput = {bits: string};
export type AddressOutput = AddressInput;
export type AssetIdInput = {bits: string};
export type AssetIdOutput = AssetIdInput;
export type ContractIdInput = {bits: string};
export type ContractIdOutput = ContractIdInput;

export type SwapExactOutputScriptLoaderInputs = [
  amount_out: BigNumberish,
  asset_out: AssetIdInput,
  amount_in_max: BigNumberish,
  pools: Vec<[AssetIdInput, AssetIdInput, boolean]>,
  recipient: IdentityInput,
  deadline: BigNumberish,
];
export type SwapExactOutputScriptLoaderOutput = Vec<[BN, AssetIdOutput]>;
// ...
const bytecode = decompressBytecode(
  "H4sIAAAAAAAAA5NyMGAIcGQwkHIJYNjlycBg5MDSCOQrxALZQJoDyG9ScBVmCHIVYPFyYWBwSq/euVdsBd8z59qYxeUau9U4+VfE7b25+Kygwpcnl6xLGMCAqQNC4wf/gcAlNTElJzMvVaEgsbg4NYUhKL+0JLXISsHTL8zRx9MlPsAxxAOkNi01tZgYMwvy83Pic1NLElMSSxKB/AAgXyEvv0ShoCi1ODUP6j5cQF3AtSI5NTUlMy9dITOvoLREITE3vxSkq7g8sQDmZhA4AwYMTBAeA2OP6mmzNje9WVA+IcCIS4L3wbbNy1MI6meGUPbAGAHTPlA6BkLbQP1pPQFCG16B0DosEFr7CYTWmAHzN4SWnwKhBZZAaWg8CkRAaRMoDVXP/wBK20BpqDjfHSi9A0LzbIHSUP3cUPdwa0Borg9Q+gKUhqrngrqHE6qPE0IxcELdxXEEQrND7WdLgNCsC6A0CwDJwp/j4AIAAA=="
);

export class SwapExactOutputScriptLoader extends Script<
  SwapExactOutputScriptLoaderInputs,
  SwapExactOutputScriptLoaderOutput
> {
  static readonly abi = abi;
  static readonly bytecode = bytecode;

  constructor(wallet: Account) {
    super(bytecode, abi, wallet);
  }
}